refactor commands folders and location of files
