+++ packages/deprecation-crawler/src/lib/cli/model.ts
@@ -1,0 +1,7 @@
+import { CommandModule } from 'yargs';
+
+export interface YargsCommandObject {
+  command: string | ReadonlyArray<string>;
+  description: string;
+  module: CommandModule;
+}
