+++ packages/deprecation-crawler/src/lib/index.ts
@@ -1,0 +1,11 @@
+import { isCrawlerModeCi } from './utils';
+import { guardAgainstDirtyRepo } from './tasks/guard-against-dirty-repository';
+import { runCli } from './cli';
+import { COMMANDS, OPTIONS } from './cli-settings';
+
+(async () => {
+  if (isCrawlerModeCi()) {
+    await guardAgainstDirtyRepo();
+  }
+  runCli({ commands: COMMANDS, options: OPTIONS });
+})();
